feat[delete_task_by_id_repository.py]: introduce edge case where the desired id do not exist in the database

+++ src/todo_clean/layer3/repository/delete_task_by_id_repository.py
@@ -20,9 +20,18 @@ class DeleteTaskByIdSqliteRepository(IDeleteTaskByIdRepository):
 
     def delete_task_by_id(self, id_: int) -> bool:
         with closing(self.conn.cursor()) as cursor:
+            if (
+                cursor.execute(
+                    "SELECT * FROM tasks WHERE id = ?", (id_,)
+                ).fetchone()
+                is None
+            ):
+                return False
             try:
                 cursor.execute("DELETE FROM tasks WHERE id = ?", (id_,))
                 self.conn.commit()
             except sqlite3.OperationalError:
+                return False
+            except sqlite3.InterfaceError:
                 return False
         return True
